Rename aux.py to auxiliary.py to make pynn works on Windows machines.
