+++ pyNN/hardware/auxiliary.py
